metrics create report stages refactored
