+++ src/main/kotlin/io/github/janbarari/gradle/analytics/metric/totalbuild/report/CreateTotalBuildReportStage.kt
@@ -20,30 +20,24 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package io.github.janbarari.gradle.analytics.metric.execution.report
+package io.github.janbarari.gradle.analytics.metric.totalbuild.report
 
 import io.github.janbarari.gradle.analytics.domain.model.metric.BuildMetric
-import io.github.janbarari.gradle.analytics.domain.model.ChartPoint
-import io.github.janbarari.gradle.analytics.domain.model.report.ExecutionReport
 import io.github.janbarari.gradle.analytics.domain.model.report.Report
+import io.github.janbarari.gradle.analytics.domain.model.report.TotalBuildReport
 import io.github.janbarari.gradle.core.Stage
-import io.github.janbarari.gradle.core.Triple
-import io.github.janbarari.gradle.extension.ensureNotNull
 import io.github.janbarari.gradle.extension.isBiggerEquals
 import io.github.janbarari.gradle.extension.isNotNull
-import io.github.janbarari.gradle.extension.isNull
 import io.github.janbarari.gradle.extension.mapToChartPoints
 import io.github.janbarari.gradle.extension.mapToTimespanChartPoints
 import io.github.janbarari.gradle.extension.maxValue
 import io.github.janbarari.gradle.extension.minValue
 import io.github.janbarari.gradle.extension.minimize
 import io.github.janbarari.gradle.extension.whenEmpty
-import io.github.janbarari.gradle.utils.DateTimeUtils
-import io.github.janbarari.gradle.utils.MathUtils
 
-class CreateExecutionReportStage(
+class CreateTotalBuildReportStage(
     private val metrics: List<BuildMetric>
-) : Stage<Report, Report> {
+): Stage<Report, Report> {
 
     companion object {
         private const val SKIP_METRIC_THRESHOLD = 50L
@@ -52,9 +46,9 @@ class CreateExecutionReportStage(
 
     override suspend fun process(report: Report): Report {
         val chartPoints = metrics.filter { metric ->
-            metric.executionMetric.isNotNull() &&
-                    metric.executionMetric?.average.isNotNull() &&
-                            metric.executionMetric?.average?.isBiggerEquals(SKIP_METRIC_THRESHOLD) ?: false
+            metric.totalBuildMetric.isNotNull() &&
+                    metric.totalBuildMetric?.average.isNotNull() &&
+                            metric.totalBuildMetric?.average?.isBiggerEquals(SKIP_METRIC_THRESHOLD) ?: false
         }.mapToTimespanChartPoints()
             .minimize(CHART_MAX_COLUMNS)
             .mapToChartPoints()
@@ -63,11 +57,12 @@ class CreateExecutionReportStage(
             }
 
         return report.apply {
-            executionReport = ExecutionReport(
+            totalBuildReport = TotalBuildReport(
                 values = chartPoints,
-                maxValue = chartPoints.maxValue(),
+                maxValue= chartPoints.maxValue(),
                 minValue = chartPoints.minValue()
             )
         }
     }
+
 }
